Store serena toggle persistently on player
Previously the toggles done by serena would not persist through a server
restart, as it was purely maintained in an in-memory queue cache.

As this was rather inconvinient for the players, this commit now offers
a new implementation for the player toggle register, based on the
persistent data container api provided by spigot, which stores the
toggle value of serena directly on the player nbt tags.

New version: 1.0.4-SNAPSHOT